cleaned up imports, removed cyclic dependencies
