+++ client/python/libs/sklearn/events/Event.py
@@ -1,0 +1,11 @@
+import numpy as np
+import pandas as pd
+import sys
+sys.path.append('./thrift/gen-py')
+import modeldb.ttypes as modeldb_types
+
+class Event:
+    def __init__(self, experimentRun):
+        self.experimentRun = experimentRun
+    def sync(self, syncer):
+        pass
